requirements.txt
streamlit
pandas
langchain
faiss-cpu
openai
python-dotenv
fastapi[all]
pydantic==1.*
tiktoken